requisito 10 sendo feito

+++ src/components/ProductCard.js
@@ -3,13 +3,34 @@ import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
 
 class ProductCard extends Component {
-  handleAddProductToCart(product) {
+  constructor() {
+    super();
+    this.state = {
+      quantity: 1,
+      quantity2: 0,
+    };
+  }
+
+  async handleAddProductCart (product) {
     const items = this.getCartItems();
     if (items) {
-      const newItemsList = [...items, product];
-      localStorage.setItem('carrinho', JSON.stringify(newItemsList));
+      const itemExist = items.some((item) => item.id === product.id);
+      if (itemExist) {
+        const index = items.map(object => object.id).indexOf(product.id);
+        const prevQuantity = items[index].quantity
+        items[index] = { ...items[index], quantity: prevQuantity +1 }
+        localStorage.setItem('carrinho', JSON.stringify(items));
+      } else {
+        const newItemsList = [...items, product];
+        localStorage.setItem('carrinho', JSON.stringify(newItemsList));
+        this.setState((estadoAnterior) => ({
+          quantity: estadoAnterior.quantity + 1,
+          quantity2: 1,
+        }) )
+      }
     } else {
       localStorage.setItem('carrinho', JSON.stringify([product]));
+      this.setState({ quantity2: 1 })
     }
   }
 
@@ -20,7 +41,7 @@ class ProductCard extends Component {
 
   render() {
     const { product, inHome = false } = this.props;
-    const { id, price, thumbnail, title } = product;
+    const { id, price, thumbnail, title, quantity } = product;
     return (
       <div className="product-card" data-testid="product">
         <Link
@@ -31,13 +52,13 @@ class ProductCard extends Component {
           <img src={ thumbnail } alt={ title } />
           <h3 data-testid="shopping-cart-product-name">{ title }</h3>
           <p>{ price }</p>
-          <p data-testid="shopping-cart-product-quantity">1</p>
+          <p data-testid="shopping-cart-product-quantity">{ quantity }</p>
         </Link>
         { inHome && (
           <button
             data-testid="product-add-to-cart"
             type="button"
-            onClick={ () => this.handleAddProductToCart({ id, price, thumbnail, title }) }
+            onClick={ () => this.handleAddProductCart({ id, price, thumbnail, title, quantity: quantity }) }
           >
             Adicionar ao carrinho
           </button>
@@ -58,3 +79,16 @@ ProductCard.propTypes = {
 };
 
 export default ProductCard;
+
+
+
+   // const index = items.indexOf(product.id === items.id);
+        // const newItemsList = [...items, product];
+        // localStorage.setItem('carrinho', JSON.stringify(newItemsList));
+        // console.log(index)
+        // // const teste = items.map((item) => {
+        // //   if (item.id === product.id) {
+        // //     return Object.assign(item, product);
+        // //   }
+        // // })
+        // // console.log(teste)
